Split slash parity gates
