+++ docs/DECISIONS.md
@@ -288,3 +288,4 @@
 - Streaming shell command lifecycle lives in `ShellStreamingProcessRunner`; `ShellToolExecutor` remains the public shell facade for blocking, cancellable, and streaming entry points while the runner owns async stdout/stderr readers, timeout termination, cancellation, and final streamed `ToolResult` emission.
 - Top-bar parity gates live in `ParityTopBarGateTests`, not the broad `ParityGateTests` catch-all. Presentation semantics, lifecycle status labels, runtime/auth labels, model-catalog projection, and top-bar surface contract gates should stay together so top-bar UX can evolve without inflating the main architecture-gate file.
 - Model picker search semantics live in `ModelCategorySearchFilter`, not `TopBarSurface` or the SwiftUI picker. `TopBarSurface.filteredModelCategories(matching:)` remains the compatibility entry point, while query normalization, Favorites/Recent visibility, metadata haystack construction, and State-row search behavior stay directly testable beside the filter.
+- Slash parser architecture gates live in `ParitySlashGateTests`, not the broad `ParityGateTests` catch-all. PR, project, terminal, and mode delegation checks should stay together so slash-command parser extraction work does not keep inflating the main architecture-gate file.
